+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKB0P285P7QMP52ZSKFHXY2V"
+        "01KKBJ1G37AH6M7F3SAHYVRP8E"
     )
     r"""The unique identifier of the data source"""