+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -96,7 +96,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-03-10T04:39:20.491Z")
+    created: Optional[datetime] = parse_datetime("2026-03-10T09:42:40.419Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")