+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKB0P28GFK4C12DMDT32H8PQ"
+        "01KKBJ1G399AH11Q0Y6F3PH2ZF"
     )
     r"""The unique identifier of the data source"""